pyproject.toml
[tool.poetry]
name = "whatsapp_tomticket"
version = "0.1.0"
description = "A WhatsApp chatbot to interact with Tomticket"
authors = ["Paulo Vitor paulovsp16@hotmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
requests = "^2.26.0"
twilio = "^6.63.2"
python-dotenv = "^0.19.0"
lxml = "^4.6.3"
beautifulsoup4 = "^4.9.3"
google-cloud-dialogflow = "^2.7.1"
google-api-core = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.25.0"
flask-shell-ipython = "^0.4.1"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
